connect productPage frontEnd

+++ src/main/java/lk/ijse/helloshoesbackend/service/iml/GenderServiceImpl.java
@@ -2,6 +2,8 @@ package lk.ijse.helloshoesbackend.service.iml;
 
 import jakarta.transaction.Transactional;
 import lk.ijse.helloshoesbackend.dto.GenderDTO;
+import lk.ijse.helloshoesbackend.exception.DuplicateException;
+import lk.ijse.helloshoesbackend.exception.NotFoundException;
 import lk.ijse.helloshoesbackend.repository.GenderDao;
 import lk.ijse.helloshoesbackend.service.GenderService;
 import lk.ijse.helloshoesbackend.util.Mapping;
@@ -19,6 +21,8 @@ public class GenderServiceImpl implements GenderService {
     private final Mapping mapping;
     @Override
     public void saveGender(GenderDTO genderDTO) {
+        System.out.println(genderDTO.getGenderCode()+"=========================");
+        if (genderDao.existsById(genderDTO.getGenderCode())) throw new DuplicateException("Gender id Duplicate");
         genderDao.save(mapping.toGenderEntity(genderDTO));
     }
 
@@ -29,12 +33,13 @@ public class GenderServiceImpl implements GenderService {
 
     @Override
     public void updateGender(String id, GenderDTO genderDTO) {
+        if(!genderDao.existsById(id)) throw new NotFoundException("Gender Not Found");
         genderDao.save(mapping.toGenderEntity(genderDTO));
     }
 
     @Override
-    public boolean deleteGender(String id) {
+    public void deleteGender(String id) {
+        if (!genderDao.existsById(id)) throw new NotFoundException("Gender Not Found");
         genderDao.deleteById(id);
-        return true;
     }
 }
